+++ Assets/Scripts/Sound/Music.cs
@@ -1,0 +1,14 @@
+using System;
+using UnityEngine;
+
+[Serializable]
+public class Music : SoundBase
+{
+    public MusicName name;
+
+    public override void Init(GameObject gameObject)
+    {
+        base.Init(gameObject);
+        source.loop = true;
+    }
+}
