+++ Assets/Scripts/Sound/Music.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 466f3ec462d189c43afd511b4fa8a985
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
