+++ Assets/Scripts/Sound/Sound.cs
@@ -1,0 +1,14 @@
+using System;
+using UnityEngine;
+
+[Serializable]
+public class Sound : SoundBase
+{
+    public SoundName name;
+
+    public override void Init(GameObject gameObject)
+    {
+        base.Init(gameObject);
+        source.loop = false;
+    }
+}
