+++ Assets/Scripts/Sound/SoundBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8aa05f5d211810143aca01e48fe62174
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
