+++ Assets/Scripts/Sound/SoundException.cs
@@ -1,0 +1,6 @@
+using System;
+
+public class SoundException : Exception
+{
+    public SoundException(string msg) : base(msg) { }
+}
