+++ Assets/Scripts/Sound/SoundException.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ab1dc4fab9ab2045bf0e0d8b3e91a65
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
