+++ Assets/Scripts/Sound/SoundName.cs
@@ -1,0 +1,10 @@
+public enum SoundName
+{
+    PlayerTakeDamage, PlayerDeath, PlayerWalk, PlayerJump, PlayerThrow, PlayerMelee,
+    EnemyTakeDamage, EnemyDeath, EnemyWalk, EnemyJump, EnemyShoot, EnemyMelee, //another sound for dark night, dark magician, dark knight etc?
+}
+
+public enum MusicName
+{
+    MainMenuTheme, Level1Theme, Level2Theme, Level3Theme
+}
