+++ Assets/Scripts/Sound/SoundName.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85d667fc60fa73c4b9f6d5390025da09
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
